work towards enabling the debug IPC service on Linux
