+++ rust/gui-client/src-tauri/deb_files/firezone-client-ipc.service
@@ -42,7 +42,7 @@ Environment="LOG_DIR=/var/log/dev.firezone.client"
 Environment="RUST_LOG=info"
 EnvironmentFile="/etc/default/firezone-client-ipc"
 
-ExecStart=firezone-client-ipc
+ExecStart=firezone-client-ipc ipc-service
 Type=notify
 # Unfortunately we may need root to control DNS
 User=root
